fix: report a fill's own move, and read a historical size as stated
Two faults of the same kind, one in what this branch added and one it
had not reached yet.

A fill moves the holding and the broker does not restate holdings when
an order fills, so the fill is the only account of it. The fill updated
the holding without recording that it had moved, and the real-time
report skipped the one change a caller watching positions is most
likely watching for.

A historical tick size crosses as text, not as a number, because a size
can be a fraction of a share. Read as a whole number, a fractional print
came back as a print of nothing. The three historical tick sizes are
decimals now, like the price beside them.

+++ src/python/compat/client/dispatch.rs
@@ -742,7 +742,7 @@ impl EClient {
                         time: at(&t.time).or_else(|| { dropped.set(dropped.get() + 1); None })?,
                         tick_attrib_last: Default::default(),
                         price: t.price,
-                        size: t.size as f64,
+                        size: t.size,
                         exchange: t.exchange.clone(),
                         special_conditions: t.special_conditions.clone(),
                     })).collect();
@@ -755,8 +755,8 @@ impl EClient {
                         tick_attrib_bid_ask: Default::default(),
                         price_bid: t.bid_price,
                         price_ask: t.ask_price,
-                        size_bid: t.bid_size as f64,
-                        size_ask: t.ask_size as f64,
+                        size_bid: t.bid_size,
+                        size_ask: t.ask_size,
                     })).collect();
                     let list = pyo3::types::PyList::new(py, py_ticks)?;
                     call_wrapper!(self.wrapper, py, "historical_ticks_bid_ask", (req_id as i64, list, done));
